requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
cryptography>=42.0.0
base58>=2.1.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
